Remove scenario tests completely and move all tests cases to integration tests
