+++ .github/workflows/test.yml
@@ -184,13 +184,6 @@ jobs:
           command: build
           args: --tests --all-features ${{ env.CARGO_FLAGS }}
 
-      - name: Build scenario tests
-        uses: actions-rs/cargo@v1
-        timeout-minutes: 30
-        with:
-          command: build
-          args: -p jormungandr-scenario-tests ${{ env.CARGO_FLAGS }}
-
       - name: Run tests
         uses: actions-rs/cargo@v1
         timeout-minutes: 90
