+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
   "modules/blockchain",
   "testing/jormungandr-testing-utils",
   "testing/jormungandr-integration-tests",
-  "testing/jormungandr-scenario-tests",
   "testing/loki",
   "testing/mjolnir",
   "testing/hersir",
