+++ testing/jormungandr-integration-tests/Cargo.toml
@@ -19,6 +19,7 @@ chain-impl-mockchain = { git = "https://github.com/input-output-hk/chain-libs.gi
 chain-time      = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
 chain-vote      = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
 jormungandr-lib = { path = "../../jormungandr-lib" }
+hersir = { path = "../hersir" }
 jormungandr-testing-utils = { path = "../jormungandr-testing-utils" }
 jortestkit = { git = "https://github.com/input-output-hk/jortestkit.git", branch = "master" }
 rand = "0.8"
@@ -44,6 +45,7 @@ json = "0.12.4"
 multiaddr = { package = "parity-multiaddr", version = "0.11" }
 rstest = "0.10.0"
 tracing = "0.1"
+function_name = "0.2.0"
 
 [dependencies.reqwest]
 version = "0.11"
@@ -52,6 +54,9 @@ features = ["blocking", "json", "rustls-tls"]
 
 [features]
 testnet = []
+network = []
+network-non-functional = []
 sanity-non-functional = []
 soak-non-functional = []
+
 explorer-schema-gen = []
