+++ testing/jormungandr-integration-tests/src/networking/cross_version/mod.rs
@@ -12,6 +12,3 @@ const ALICE: &str = "ALICE";
 const BOB: &str = "BOB";
 const CLARICE: &str = "CLARICE";
 const DAVID: &str = "DAVID";
-
-pub use disruption::*;
-pub use fragment_propagation::*;
