+++ testing/jormungandr-integration-tests/src/networking/mod.rs
@@ -1,4 +1,12 @@
+pub mod bft;
+pub mod big;
 pub mod communication;
+pub mod cross_version;
+pub mod explorer;
+pub mod leadership_log;
 pub mod p2p;
+pub mod stake_pool;
 pub mod stats;
 pub mod testnet;
+pub mod topology;
+pub mod utils;
