+++ testing/jormungandr-integration-tests/src/networking/p2p/mod.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)]
-
 pub mod connections;
 pub mod stats;
