+++ testing/jormungandr-integration-tests/src/networking/stake_pool/mod.rs
