+++ testing/jormungandr-integration-tests/src/non_functional/mod.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "network-non-functional")]
+pub mod network;
+
 #[cfg(feature = "sanity-non-functional")]
 pub mod compatibility;
 /*
