+++ testing/jormungandr-integration-tests/src/non_functional/network/desync.rs
@@ -1,4 +1,4 @@
-use crate::test::non_functional::*;
+use crate::non_functional::network::*;
 use chain_impl_mockchain::chaintypes::ConsensusVersion;
 use function_name::named;
 use jormungandr_testing_utils::testing::network::blockchain::BlockchainBuilder;
@@ -9,6 +9,7 @@ use jormungandr_testing_utils::testing::network::SpawnParams;
 use jormungandr_testing_utils::testing::network::Topology;
 use jormungandr_testing_utils::testing::FragmentSender;
 use jormungandr_testing_utils::{testing::network::FaketimeConfig, wallet::Wallet};
+
 #[named]
 pub fn bft_forks() {
     let n_transactions = 5;
