+++ testing/jormungandr-integration-tests/src/non_functional/network/disruption.rs
@@ -1,5 +1,5 @@
-use crate::test::non_functional::*;
-use crate::test::utils;
+use crate::networking::utils;
+use crate::non_functional::network::*;
 use jormungandr_testing_utils::testing::network::builder::NetworkBuilder;
 use jormungandr_testing_utils::testing::network::wallet::template::builder::WalletTemplateBuilder;
 use jormungandr_testing_utils::testing::network::Node;
