--- testing/jormungandr-scenario-tests/Cargo.toml
@@ -1,49 +1,0 @@
-[package]
-name = "jormungandr-scenario-tests"
-version = "0.13.0"
-authors = [ "dev@iohk.io" ]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-custom_debug = "0.5"
-dialoguer = "0.8.0"
-assert_fs = "1.0"
-chain-core           = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
-chain-crypto         = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = [ "property-test-api" ] }
-chain-addr           = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = [ "property-test-api" ] }
-chain-impl-mockchain = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master", features = [ "property-test-api" ] }
-chain-time           = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
-jormungandr-testing-utils = { path = "../jormungandr-testing-utils" }
-jormungandr-lib = { path = "../../jormungandr-lib" }
-hersir = { path = "../hersir" }
-jortestkit = { git = "https://github.com/input-output-hk/jortestkit.git", branch = "master" }
-tokio = { version = "1.4", features = ["macros","rt","rt-multi-thread"] }
-poldercast = "1.2"
-rand = "0.8"
-rand_core = "0.6"
-rand_chacha = "0.3"
-json = "0.12.4"
-function_name = "0.2.0"
-chrono = "0.4.19"
-regex = "1.4"
-rstest = "0.11.0"
-serde = "1.0"
-serde_derive = "1.0"
-serde_json = "1.0"
-serde_yaml = "0.8"
-structopt = "0.3"
-thiserror = "1.0"
-hex = "0.4"
-futures = "0.3.18"
-console = "0.15"
-yaml-rust = "0.4.4"
-indicatif = "0.15"
-lazy_static = "1"
-multiaddr = { package = "parity-multiaddr", version = "0.11" }
-
-[dependencies.reqwest]
-version = "0.11"
-default-features = false
-features = ["blocking", "rustls-tls"]
