--- testing/jormungandr-scenario-tests/src/test/features/mod.rs
@@ -1,4 +1,0 @@
-pub mod explorer;
-pub mod leadership_log;
-pub mod p2p;
-pub mod stake_pool;
