--- testing/jormungandr-scenario-tests/src/test/network/mod.rs
@@ -1,3 +1,0 @@
-pub mod bft;
-pub mod real;
-pub mod topology;
