--- testing/jormungandr-scenario-tests/src/test/network/topology/mod.rs
@@ -1,1 +1,0 @@
-pub mod scenarios;
